Unfork LOGISTRATION feature flag
We originally forked this in order to easily toggle use of the combined
logistration form. At the time, we were still using the old-style
login/register pages. As time went by, we began to enable LOGISTRATION
on deployments one-by-one. Now that we're rolling the change out to
SUclass, we no longer need the antiquated forms.

We (the OpenEdX community) probably no longer need this flag since I
suspect all deployments should be using the combined form, but regardless,
this changes brings us back in line with edx.org.