Remove deprecated ~/.ssh/config helper
This was a nice idea, but it can quickly fall out of date.
At that point, it's up to the user to remember to refresh their config.

Now that we're using internal domain names, it's easy to set up the
`~/.ssh/config` file manually.
Instructions can be found in the deployment README files.