Enable bottom of progress page
We had feature-flagged this some time ago, due to the unbearably slow
loading times; pages would often time out before ever loading.

Now that edX has invested in increasing the performance of this page,
we've confirmed that most courses now load either immediately or within
a few seconds, as opposed to the previous 30-60+ second load times.

We'll monitor the results with this re-enabled and once we've confirmed
everything is working well, we can look to remove the feature flag
entirely in a future sprint.