Omit successfull `/heartbeat` requests from logs
This has long been a frustration of mine.
It's a pain to be `tail`ing logs and have the screen fill up with
heartbeat checks every few seconds.

I was prompted to revisit this following the recent issue with log files
taking up a significant amount of disk space.

Note: only successful attempts are excluded from the logs.
Should an error occur, it will still be logged to the error log file.