Fix cluster role's optional include
It looks like this hasn't actually been working,
because the `tasks` section isn't executed prior to the `roles`.

By just splitting this out into two separate plays,
we can ensure they are executed in the proper order.
And since the former play is simply an include,
it doesn't rely on any side-effects of the latter.